% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stanirs-package.R
\docType{package}
\name{stanirs-package}
\alias{stanirs-package}
\alias{stanirs}
\title{Package for working with fNIRS files}
\description{
Working with fNIRS files in Stanford's Center for Interdisciplinary Brain Sciences.

TODO: consider adding links to funders.
}
\note{
The most recent development version is available through \href{https://github.com/OuhscBbmc/referral-exposure}{GitHub} by
running.

\code{devtools::install_github('cibsr/stanirs')}
(make sure \href{https://cran.r-project.org/package=devtools}{devtools} is already installed).
If you're having trouble with the package, please install the development version.  If this doesn't solve
your problem, please create a \href{https://github.com/cibsr/stanirs/issues}{new issue}, or email Andrey.
}
\examples{
\dontrun{
# Install/update ReferralExposure with the development version from GitHub
#install.packages('devtools') #Uncomment if `devtools` isn't installed already.
devtools::install_github('cibsr/stanirs')
}
}
